requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
psycopg2-binary==2.9.7
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
drf-spectacular==0.26.5 